Removed tester-specific interfaces and combined into newstyle single one
