+++ .travis.yml
@@ -4,6 +4,7 @@ rvm:
   - 2.2.2
 script:
   - bundle exec origen -v
+  - bundle exec origen t dut
   - bundle exec origen test -c
   - bundle exec origen web compile --no-serve
   - bundle exec origen lint --no-correct
