+++ config/boot.rb
@@ -7,9 +7,6 @@ require "origen_testers/test/block.rb"
 require "origen_testers/test/dut2.rb"
 require "origen_testers/test/nvm.rb"
 
-require "origen_testers/test/j750_interface"
-require "origen_testers/test/ultraflex_interface"
-require "origen_testers/test/j750_hpt_interface"
-require "origen_testers/test/v93k_interface"
+require "origen_testers/test/interface"
 require "origen_testers/test/basic_interface"
 require "origen_testers/test/custom_test_interface"
