--- lib/origen_testers/test/j750_base_interface.rb
@@ -1,145 +1,0 @@
-module OrigenTesters
-  module Test
-    class J750BaseInterface
-      # Options passed to Flow.create and Library.create will be passed in here, use as
-      # desired to configure your interface
-      def initialize(options = {})
-      end
-
-      def log(msg)
-        flow.logprint(msg)
-      end
-
-      def func(name, options = {})
-        options = {
-          duration: :static
-        }.merge(options)
-
-        block_loop(name, options) do |block, i, group|
-          ins = test_instances.functional(name)
-          ins.set_wait_flags(:a) if options[:duration] == :dynamic
-          ins.pin_levels = options.delete(:pin_levels) if options[:pin_levels]
-          if group
-            pname = "#{name}_b#{i}_pset"
-            patsets.add(pname, [{ pattern: "#{name}_b#{i}.PAT" },
-                                { pattern: 'nvm_global_subs.PAT', start_label: 'subr' }])
-            ins.pattern = pname
-            flow.test(group, options) if i == 0
-          else
-            pname = "#{name}_pset"
-            patsets.add(pname, [{ pattern: "#{name}.PAT" },
-                                { pattern: 'nvm_global_subs.PAT', start_label: 'subr' }])
-            ins.pattern = pname
-            if options[:cz_setup]
-              flow.cz(ins, options[:cz_setup], options)
-            else
-              flow.test(ins, options)
-            end
-          end
-        end
-      end
-
-      def block_loop(name, options)
-        if options[:by_block]
-          test_instances.group do |group|
-            group.name = name
-            $dut.blocks.each_with_index do |block, i|
-              yield block, i, group
-            end
-          end
-        else
-          yield
-        end
-      end
-
-      def por(options = {})
-        options = {
-          instance_not_available: true
-        }.merge(options)
-        flow.test('por_ins', options)
-      end
-
-      def para(name, options = {})
-        options = {
-          high_voltage: false
-        }.merge(options)
-        if options.delete(:high_voltage)
-          ins = test_instances.bpmu(name)
-        else
-          ins = test_instances.ppmu(name)
-        end
-        ins.dc_category = 'NVM_PARA'
-        flow.test(ins, options)
-        patsets.add("#{name}_pset", pattern: "#{name}.PAT")
-      end
-
-      # OR 2 IDS together into 1 flag
-      def or_ids(options = {})
-        flow.or_flags(options[:id1], options[:id2], options)
-      end
-
-      def nop(options = {})
-        flow.nop options
-      end
-
-      def mto_memory(name, options = {})
-        options = {
-          duration: :static
-        }.merge(options)
-
-        block_loop(name, options) do |block, i, group|
-          ins = test_instances.mto_memory(name)
-          ins.set_wait_flags(:a) if options[:duration] == :dynamic
-          ins.pin_levels = options.delete(:pin_levels) if options[:pin_levels]
-          if group
-            pname = "#{name}_b#{i}_pset"
-            patsets.add(pname, [{ pattern: "#{name}_b#{i}.PAT" },
-                                { pattern: 'nvm_global_subs.PAT', start_label: 'subr' }])
-            ins.pattern = pname
-            flow.test(group, options) if i == 0
-          else
-            pname = "#{name}_pset"
-            patsets.add(pname, [{ pattern: "#{name}.PAT" },
-                                { pattern: 'nvm_global_subs.PAT', start_label: 'subr' }])
-            ins.pattern = pname
-            if options[:cz_setup]
-              flow.cz(ins, options[:cz_setup], options)
-            else
-              flow.test(ins, options)
-            end
-          end
-        end
-
-        def meas(name, options = {})
-          options = {
-            duration: :static
-          }.merge(options)
-
-          name = "meas_#{name}" unless name.to_s =~ /meas/
-
-          if options[:pins] == :hi_v
-            ins = test_instances.board_pmu(name)
-          elsif options[:pins] == :power
-            ins = test_instances.powersupply(name)
-          else
-            ins = test_instances.pin_pmu(name)
-          end
-          ins.set_wait_flags(:a) if options[:duration] == :dynamic
-          ins.pin_levels = options.delete(:pin_levels) if options[:pin_levels]
-          ins.lo_limit = options[:lo_limit]
-          ins.hi_limit = options[:hi_limit]
-
-          pname = "#{name}_pset"
-          patsets.add(pname, [{ pattern: "#{name}.PAT" },
-                              { pattern: 'nvm_global_subs.PAT', start_label: 'subr' }])
-          ins.pattern = pname
-          if options[:cz_setup]
-            flow.cz(ins, options[:cz_setup], options)
-          else
-            flow.test(ins, options)
-          end
-        end
-      end
-    end
-  end
-end
